Modified get user name function, moved csv file creation to reporting.py
